@@ -10,49 +10,54 @@
   "author": "",
   "license": "ISC",
   "devDependencies": {
-    "@types/react": "^15.0.28",
-    "@types/react-dom": "^15.5.0",
+    "@types/googlemaps": "^3.29.2",
+    "@types/node": "^8.0.46",
+    "@types/react": "^16.0.16",
+    "flatpickr": "^4.0.6",
+    "@types/react-dom": "^16.0.2",
     "assets-webpack-plugin": "^3.5.1",
-    "awesome-typescript-loader": "^3.1.3",
-    "babel-core": "^6.24.1",
-    "babel-loader": "^6.4.1",
-    "babel-polyfill": "^6.23.0",
-    "babel-preset-env": "^1.4.0",
+    "awesome-typescript-loader": "^3.2.3",
+    "babel-core": "^6.26.0",
+    "babel-loader": "^7.1.2",
+    "babel-polyfill": "^6.26.0",
+    "babel-preset-env": "^1.6.1",
     "babel-preset-es2015": "^6.24.1",
     "babel-preset-react": "^6.24.1",
     "babel-preset-stage-2": "^6.24.1",
-    "copy-webpack-plugin": "^4.0.1",
-    "css-loader": "^0.28.0",
+    "copy-webpack-plugin": "^4.2.0",
+    "css-loader": "^0.28.7",
     "cssnano": "^3.10.0",
-    "es6-promise": "^4.1.0",
+    "es6-promise": "^4.1.1",
     "expose-loader": "^0.7.3",
-    "extract-text-webpack-plugin": "^2.1.0",
-    "file-loader": "^0.11.1",
+    "extract-text-webpack-plugin": "^3.0.1",
+    "file-loader": "^1.1.5",
     "isomorphic-fetch": "^2.2.1",
-    "node-sass": "^4.5.2",
-    "normalizr": "^3.2.2",
-    "optimize-css-assets-webpack-plugin": "^1.3.1",
-    "react": "^15.5.4",
-    "react-dom": "^15.5.4",
-    "react-redux": "^5.0.4",
-    "redux": "^3.6.0",
-    "redux-devtools": "^3.3.2",
+    "moment": "^2.19.1",
+    "node-sass": "^4.5.3",
+    "normalizr": "^3.2.4",
+    "optimize-css-assets-webpack-plugin": "^3.2.0",
+    "react": "^16.0.0",
+    "react-dom": "^16.0.0",
+    "react-redux": "^5.0.6",
+    "redux": "^3.7.2",
+    "redux-devtools": "^3.4.0",
     "redux-thunk": "^2.2.0",
-    "sass-loader": "^6.0.3",
-    "source-map-loader": "^0.2.1",
-    "style-loader": "^0.16.1",
-    "typescript": "^2.3.4",
-    "url-join": "^2.0.1",
-    "url-loader": "^0.5.8",
-    "webpack": "^2.4.1",
+    "sass-loader": "^6.0.6",
+    "socketcluster-client": "^8.0.1",
+    "source-map-loader": "^0.2.2",
+    "style-loader": "^0.19.0",
+    "typescript": "^2.5.3",
+    "url-join": "^2.0.2",
+    "url-loader": "^0.6.2",
+    "webpack": "^3.8.1",
     "webpack-merge": "^4.1.0"
   },
   "dependencies": {
     "@types/bootstrap": "^3.3.36",
-    "@types/flatpickr": "^3.0.2",
-    "@types/jquery": "^3.2.12",
-    "@types/jquery.datatables": "^1.10.36",
-    "@types/moment": "^2.13.0",
-    "@types/moment-duration-format": "^1.3.7"
+    "@types/datatables.net": "^1.10.6",
+    "@types/jquery": "^3.2.15",
+    "@types/moment-duration-format": "^1.3.7",
+    "install": "^0.10.1",
+    "npm": "^5.5.1"
   }
 }
