@@ -3,11 +3,10 @@
 namespace App\Mail;
 
 use Illuminate\Bus\Queueable;
-use Illuminate\Mail\Mailable;
 use Illuminate\Queue\SerializesModels;
-use Illuminate\Contracts\Queue\ShouldQueue;
 
-class ResetPassword extends Mailable
+
+class ResetPassword extends BaseMailable
 {
     public $token;
     use Queueable, SerializesModels;
