@@ -1,1 +1,49 @@
-webpackJsonp([2],{0:function(e,t,n){"use strict";function a(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}Object.defineProperty(t,"__esModule",{value:!0});var r=function e(){a(this,e),this.timeHelper=null},c=new r;t.default=c},19:function(e,t,n){"use strict";n(7)},5:function(e,t,n){"use strict";function a(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}var r=function(){function e(e,t){for(var n=0;n<t.length;n++){var a=t[n];a.enumerable=a.enumerable||!1,a.configurable=!0,"value"in a&&(a.writable=!0),Object.defineProperty(e,a.key,a)}}return function(t,n,a){return n&&e(t.prototype,n),a&&e(t,a),t}}();Object.defineProperty(t,"__esModule",{value:!0});var c=function(){function e(){a(this,e)}return r(e,[{key:"replaceUrlParams",value:function(e,t,n){var a=$("#appId").val();if(e=e.replace("###",a),n.length>0){var r=$(t).data(n);e=e.replace("##",r)}return e}}]),e}(),s=new c;t.default=s},7:function(e,t,n){"use strict";Object.defineProperty(t,"__esModule",{value:!0});var a=n(0),r=n(5);$(document).on("click",".js-transaction-accept,.js-transaction-reject",function(){var e=a.default.ajaxUrls.moneymakerTransactionStatusUrl,t=$(this);e=r.default.replaceUrlParams(e,t,"trans-id");var n=t.data("status"),c=new FormData;c.append("status",n),$.siteAjax({url:e,method:"post",processData:!1,contentType:!1,data:c,success:function(e){"string"==typeof e&&(e=JSON.parse(e));var a=t.parent();a.html('<span class="text-center status-text" >Accepted</span>'),3===n&&a.find("span").text("Rejected")}})})}},[19]);
+webpackJsonp([3],{
+
+/***/ 20:
+/***/ (function(module, exports, __webpack_require__) {
+
+"use strict";
+
+
+__webpack_require__(21);
+
+/***/ }),
+
+/***/ 21:
+/***/ (function(module, exports, __webpack_require__) {
+
+"use strict";
+
+
+Object.defineProperty(exports, "__esModule", { value: true });
+var globals_1 = __webpack_require__(0);
+var common_1 = __webpack_require__(5);
+$(document).on('click', '.js-transaction-accept,.js-transaction-reject', function () {
+    var url = globals_1.default.ajaxUrls.moneymakerTransactionStatusUrl;
+    var button = $(this);
+    url = common_1.default.replaceUrlParams(url, button, 'trans-id');
+    var status = button.data('status');
+    var fD = new FormData();
+    fD.append('status', status);
+    $.siteAjax({
+        url: url,
+        method: 'post',
+        processData: false,
+        contentType: false,
+        data: fD,
+        success: function success(data) {
+            if (typeof data == 'string') data = JSON.parse(data);
+            var td = button.parent();
+            td.html('<span class="text-center status-text" >Accepted</span>');
+            if (status === 3) {
+                td.find('span').text('Rejected');
+            }
+        }
+    });
+});
+
+/***/ })
+
+},[20]);
+//# sourceMappingURL=transaction.js.map
