@@ -2,12 +2,7 @@
 @section('title', 'Application DashBoard')
 
 @section('content')
-    <div class="container">
-        <div class="sidenav">
-            {!! $DashboardNavbar->asUl() !!}
-        </div>
-        <div class="row">
-            <div class="col-md-offset-1 col-md-8">
+
                 <h1 style="text-align: left;">Applications</h1>
                 <hr style="width:250px;" align="left">
 
@@ -20,7 +15,5 @@
                         </a>
                     @endforeach
                 </div>
-            </div>
-        </div>
-    </div>
+
 @endsection
