@@ -14,6 +14,7 @@
 @endsection
 
 @section('content')
+    @parent
     @section('scripts')
         <script>
             $(document).ready(function(){
@@ -22,12 +23,11 @@
             });
         </script>
     @endsection
-    <div class="container">
+    <div>
         <input type="hidden" id="appId" value="{{$application->id}}"/>
         <input type="hidden" id="appSlug" value="{{$application->route_prefix}}"/>
-        <div class="col-sm-offset-1 col-sm-9">
+        <div>
             <h1 style="text-align: left;">"{{$application->name}}" Transactions</h1>
-            <hr style="width:250px;" align="left">
 
             <input type="hidden" name="tab" value="{{$tab}}"/>
             <div class="list-group">
@@ -39,27 +39,35 @@
                 <div class="tab-content">
                     <div class="tab-pane fade in active">
                         <br/><br/>
-                        <table class="table-hover table">
-                            <th>Username</th>
-                            <th>User ID</th>
-                            <th>Amount</th>
-                            @if($tab==='Pending')
-                                <th>Actions</th>
-                            @endif
+                        <table class="table-bordered table table-striped">
+                            <thead>
+                                <tr>
+                                    <th>Username</th>
+                                    <th>User ID</th>
+                                    <th>Amount</th>
+                                    @if($tab==='Pending')
+                                        <th colspan="2">Actions</th>
+                                    @endif
+                                </tr>
+                            </thead>
                         @foreach($transactions as $tran)
                             <tr>
                                 <td>{{ $tran->getUsername($tran->app_user_id) }}</td>
                                 <td>{{ $tran->app_user_id }}</td>
                                 <td>{{config('moneymaker.currency')}}{{ $tran->amount }}</td>
                                 @if($tab==='Pending')
-                                    <td class="text-center">
-                                        @if($tran->status===\App\Models\AppUserTransaction::$STATUS_PENDING)
+                                    @if($tran->status===\App\Models\AppUserTransaction::$STATUS_PENDING)
+                                        <td class="text-center">
                                             <button class="js-transaction-accept btn btn-primary" data-status="{{ \App\Models\AppUserTransaction::$STATUS_ACCEPTED }}"  data-trans-id="{{$tran->id}}">Accept</button>
+                                        </td>
+                                        <td class="text-center">
                                             <button class="js-transaction-reject btn btn-danger" data-status="{{ \App\Models\AppUserTransaction::$STATUS_REJECTED }}" data-trans-id="{{$tran->id}}">Reject</button>
-                                        @els
+                                        </td>
+                                    @else
+                                        <td class="text-center" colspan="2">
                                             <span class="text-center status-text" >{{  $tran->status_str }}</span>
-                                        @endif
-                                    </td>
+                                        </td>
+                                    @endif
                                 @endif
                             </tr>
                         @endforeach
