@@ -2,17 +2,10 @@
 
 @section('appName')
     @parent
-    {{$application->name}}
+    {{ $application->name }}
 @endsection
 
-@section('sidebar')
-    @if(Auth::check() && !empty($DashboardNavbar))
-        <div class="sidenav">
-            {!! $DashboardNavbar->asUl() !!}
-        </div>
-    @endif
-@endsection
+
 @section('styles')
     @parent
-
 @endsection
