@@ -5,52 +5,48 @@
 @endsection
 
 @section('content')
-    <div class="container">
-        <div class="row">
-            <div class="col-md-8 col-md-offset-2">
-                <div class="panel panel-default">
-                    <div class="panel-heading">Reset Password</div>
+    <div class="col-md-8 center-div">
+        <div class="panel panel-default">
+            <div class="panel-heading">Reset Password</div>
 
-                    <div class="panel-body">
+            <div class="panel-body">
 
-                        <form class="form-horizontal js-ajax-form" data-form="ajaxformsubmitter" role="form" method="POST" action="{{ route('users.reset-password-save',['token'=>$data->token]) }}">
-                            {{ csrf_field() }}
+                <form class="form-horizontal js-ajax-form" data-form="ajaxformsubmitter" role="form" method="POST" action="{{ route('users.reset-password-save',['token'=>$data->token]) }}">
+                    {{ csrf_field() }}
 
-                            <div class="form-group">
-                                <label for="email" class="col-md-4 control-label">E-Mail Address</label>
+                    <div class="form-group">
+                        <label for="email" class="col-md-4 control-label">E-Mail Address</label>
 
-                                <div class="col-md-6">
-                                    <input id="email" type="email" class="form-control" name="email" value="{{ $data->email }}" required autofocus disabled>
-                                </div>
-                            </div>
-
-                            <div class="form-group">
-                                <label for="password" class="col-md-4 control-label">Password</label>
-
-                                <div class="col-md-6">
-                                    <input id="password" type="password" class="form-control" name="password" required>
-                                    <div class="form-error" data-error="password"></div>
-                                </div>
-                            </div>
-
-                            <div class="form-group">
-                                <label for="password-confirm" class="col-md-4 control-label">Confirm Password</label>
-                                <div class="col-md-6">
-                                    <input id="password-confirm" type="password" class="form-control" name="password_confirmation" required>
-                                    <div class="form-error" data-error="password_confirmation"></div>
-                                </div>
-                            </div>
-
-                            <div class="form-group">
-                                <div class="col-md-6 col-md-offset-4">
-                                    <button type="submit" class="btn btn-primary">
-                                        Reset Password
-                                    </button>
-                                </div>
-                            </div>
-                        </form>
+                        <div class="col-md-6">
+                            <input id="email" type="email" class="form-control" name="email" value="{{ $data->email }}" required autofocus disabled>
+                        </div>
                     </div>
-                </div>
+
+                    <div class="form-group">
+                        <label for="password" class="col-md-4 control-label">Password</label>
+
+                        <div class="col-md-6">
+                            <input id="password" type="password" class="form-control" name="password" required>
+                            <div class="form-error" data-error="password"></div>
+                        </div>
+                    </div>
+
+                    <div class="form-group">
+                        <label for="password-confirm" class="col-md-4 control-label">Confirm Password</label>
+                        <div class="col-md-6">
+                            <input id="password-confirm" type="password" class="form-control" name="password_confirmation" required>
+                            <div class="form-error" data-error="password_confirmation"></div>
+                        </div>
+                    </div>
+
+                    <div class="form-group">
+                        <div class="col-md-6 col-md-offset-4">
+                            <button type="submit" class="btn btn-primary">
+                                Reset Password
+                            </button>
+                        </div>
+                    </div>
+                </form>
             </div>
         </div>
     </div>
