@@ -3,39 +3,35 @@
     Reset Password
 @endsection
 @section('content')
-    <div class="container">
-        <div class="row">
-            <div class="col-md-8 col-md-offset-2">
-                <div class="panel panel-default">
-                    <div class="panel-heading">Reset Password</div>
-                    <div class="panel-body">
-                        <form class="form-horizontal" role="form" method="POST" action="{{ route('users.send-reset-email') }}">
-                            {{ csrf_field() }}
+    <div class="col-md-8 center-div">
+        <div class="panel panel-default">
+            <div class="panel-heading">Reset Password</div>
+            <div class="panel-body">
+                <form class="form-horizontal" role="form" method="POST" action="{{ route('users.send-reset-email') }}">
+                    {{ csrf_field() }}
 
-                            <div class="form-group{{ $errors->has('email') ? ' has-error' : '' }}">
-                                <label for="email" class="col-md-4 control-label">E-Mail Address</label>
+                    <div class="form-group{{ $errors->has('email') ? ' has-error' : '' }}">
+                        <label for="email" class="col-md-4 control-label">E-Mail Address</label>
 
-                                <div class="col-md-6">
-                                    <input id="email" class="form-control" name="email" value="{{ old('email') }}" required>
+                        <div class="col-md-6">
+                            <input id="email" class="form-control" name="email" value="{{ old('email') }}" required>
 
-                                    @if ($errors->has('email'))
-                                        <span class="help-block">
-                                        <strong>{{ $errors->first('email') }}</strong>
-                                    </span>
-                                    @endif
-                                </div>
-                            </div>
-
-                            <div class="form-group">
-                                <div class="col-md-6 col-md-offset-4">
-                                    <button type="submit" class="btn btn-primary">
-                                        Send Password Reset Link
-                                    </button>
-                                </div>
-                            </div>
-                        </form>
+                            @if ($errors->has('email'))
+                                <span class="help-block">
+                                <strong>{{ $errors->first('email') }}</strong>
+                            </span>
+                            @endif
+                        </div>
                     </div>
-                </div>
+
+                    <div class="form-group">
+                        <div class="col-md-6 col-md-offset-4">
+                            <button type="submit" class="btn btn-primary">
+                                Send Password Reset Link
+                            </button>
+                        </div>
+                    </div>
+                </form>
             </div>
         </div>
     </div>
